LabelLike manager component
